# Reworked `Ask<T>` to allocate fewer objects, `await` less