--- a/src/contrib/serializers/Akka.Serialization.TestKit/AkkaSerializationSpec.cs
+++ b/src/contrib/serializers/Akka.Serialization.TestKit/AkkaSerializationSpec.cs
@@ -475,7 +475,7 @@ my-settings{
             Sys.EventStream.Subscribe(TestActor, typeof(object));
             var empty = Sys.ActorOf<EmptyActor>();
             empty.Ask("hello");
-            var f = ExpectMsg<FutureActorRef>();
+            var f = ExpectMsg<FutureActorRef<object>>();
 
 
             var message = new SomeMessage