--- a/src/core/Akka/Actor/ActorRef.cs
+++ b/src/core/Akka/Actor/ActorRef.cs
@@ -68,9 +68,9 @@ namespace Akka.Actor
     ///
     /// ActorRef implementation used for one-off tasks.
     /// </summary>
-    public class FutureActorRef : MinimalActorRef
+    public class FutureActorRef<T> : MinimalActorRef
     {
-        private readonly TaskCompletionSource<object> _result;
+        private readonly TaskCompletionSource<T> _result;
         private readonly Action _unregister;
         private readonly ActorPath _path;
 
@@ -80,7 +80,7 @@ namespace Akka.Actor
         /// <param name="result">TBD</param>
         /// <param name="unregister">TBD</param>
         /// <param name="path">TBD</param>
-        public FutureActorRef(TaskCompletionSource<object> result, Action unregister, ActorPath path)
+        public FutureActorRef(TaskCompletionSource<T> result, Action unregister, ActorPath path)
         {
             if (ActorCell.Current != null)
             {
@@ -131,7 +131,15 @@ namespace Akka.Actor
             {
                 if (Interlocked.Exchange(ref status, COMPLETED) == INITIATED)
                 {
-                    _result.TrySetResult(message);
+                    if (message is T t)
+                    {
+                        _result.TrySetResult(t);
+                    }
+                    else
+                    {
+                        _result.TrySetException(new ArgumentException(
+                            $"Received message of type [{message.GetType()}] - Ask expected message of type [{typeof(T)}]"));
+                    }
                 }
             }
         }