@Aaronontheweb 2021-06-17
Shouldn't break wire compatibility, btw - `FutureActorRef` is converted to a `RemoteActorRef` when going over the wire. 

This will show up in the test suite if it's an issue though.