@Aaronontheweb 2021-06-17
Next results - removed final `await` call and an unnecessary delegate allocation:

``` ini

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19041.1052 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.203
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  DefaultJob : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT


```
|                        Method | Iterations |      Mean |     Error |    StdDev |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |----------- |----------:|----------:|----------:|----------:|------:|------:|----------:|
| RequestResponseActorSelection |      10000 | 63.210 ms | 0.9977 ms | 0.9332 ms | 3875.0000 |     - |     - |     15 MB |
|          CreateActorSelection |      10000 |  5.502 ms | 0.0470 ms | 0.0417 ms |  953.1250 |     - |     - |      4 MB |